feat(telemetry): add swap transaction handling and migration
