+++ internal/store/onchainbtcprocessedtransaction/interface.go
@@ -13,4 +13,10 @@ type IStore interface {
 
 	// Update the status of a BTC processed transaction
 	UpdateStatus(id int, status model.BtcProcessingStatus) error
+
+	// UpdateToCompleted updates the status of a BTC processed transaction to processed
+	UpdateToCompleted(id int, btcTxHash string) error
+
+	// Get all pending BTC processed transactions
+	GetPendingTransactions() ([]model.OnchainBtcProcessedTransaction, error)
 }
